fix: build시 나는 에러 수정
