+++ app/mypage/competition/page.tsx
@@ -6,7 +6,7 @@ import { useSearchParams } from 'next/navigation';
 import Link from 'next/link';
 import { data } from '@/app/data.js';
 
-export const COMPLIST_OPTIONS = ['전체', '진행 전', '진행 중', '종료'];
+const COMPLIST_OPTIONS = ['전체', '진행 전', '진행 중', '종료'];
 
 const page = () => {
   const searchParams = useSearchParams();
