--- lib/utils/svg.d.ts
@@ -1,4 +1,0 @@
-declare module '*.svg' {
-  const content: any;
-  export default content;
-}
